Playing around with using a new playhead for the track progress slider.
Decided against it because the small image makes the user think they need to be more precise with their touches and slows down interaction. It also isn’t easy to see. The primary use for this demo app is to demonstrate color, so I wanted to have the fill on the slider set to the color instead of just the playhead, like it might be in Music.app